telegram-bot/requirements.txt
python-telegram-bot==21.0.1
requests==2.31.0
reportlab==4.0.7
Pillow==10.2.0

